requirements.txt
asgiref==3.6.0
attrs==23.1.0
black==23.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.2
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.2
idna==3.4
inflection==0.5.1
jsonschema==4.17.3
mypy-extensions==1.0.0
ndg-httpsclient==0.5.1
packaging==23.1
pathspec==0.11.1
platformdirs==3.2.0
pusher==3.3.2
pyasn1==0.5.0
pycparser==2.21
PyJWT==2.6.0
PyNaCl==1.5.0
pyOpenSSL==23.1.1
pyrsistent==0.19.3
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.28.2
six==1.16.0
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==1.26.15
